Renamed the components and pages
This is purely a cosmetic fix that will make it easier to classify components and determine what their job is
