+++ client/src/pages/Dashboard/Dashboard.page.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Container } from '@material-ui/core';
 import ProjectsContainer from '../../components/ProjectsContainer/ProjectsContainer.component';
 
-const Dashboard = () => {
+const Dashboard: React.FC = () => {
 	return (
 		<Container maxWidth='xl'>
 			<ProjectsContainer />
